birdtag-system/birdtag-system-6.6.3/thumbnail_birddetectioin/lambda_container_build/requirements.txt
# Core ML dependencies
ultralytics==8.1.2
supervision==0.18.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
torch==2.2.0
torchvision==0.17.0

# AWS dependencies
boto3==1.34.34
botocore==1.34.34

# Utility dependencies
requests==2.31.0
python-dateutil==2.8.2
matplotlib==3.8.3